Undertaken further development:
1. Product Certification
2. Product Vertification
3. Payment Withdrawal
4. Payment Generation
5. Protcol Initialization TTP

6. IN PROGRESS: Product encryption and transfer to TTP

Work done up to the step 9

+++ Protocol/src/com/stunsci/jprotocol/authority/rest/AuthorityApi.java
@@ -81,7 +81,7 @@ public class AuthorityApi {
 	 @POST
 	 @Produces("application/json")
 	 @Path("/certify/")
-	 public String certifyProduct(@FormParam("id") int id, @FormParam("merchant") String merchant) {
+	 public String certifyProduct(@FormParam("id") Long id, @FormParam("merchant") String merchant) {
 	 	EntityManager em = EmfInstanceManager.getInstance().get().createEntityManager();
 
 	 	Certificate cert = new Certificate();
